fixed Auth-RestWrite circular dependency
